save number of pexel photos

+++ imageTop2/Functions/ClearFolder.swift
@@ -1,12 +1,14 @@
 import Foundation
 
-func clearFolder(folderPath: String) {
+func clearFolder(folderPath: String, fileToKeep: String) {
     do {
         let directoryURL = URL(fileURLWithPath: folderPath)
         let fileManager = FileManager.default
         let contents = try fileManager.contentsOfDirectory(at: directoryURL, includingPropertiesForKeys: nil)
         for fileURL in contents {
-            try fileManager.removeItem(at: fileURL)
+            if fileURL.lastPathComponent != fileToKeep {
+                try fileManager.removeItem(at: fileURL)
+            }
         }
     } catch {
         // Handle the error.
